backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
channels==4.0.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
inflection==0.5.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
pillow==10.2.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
